requirements.txt
google-adk
google-adk[a2a]
python-dotenv
